movies_auth/pyproject.toml
[tool.poetry]
name = 'Auth service'
description = "Yandex practicum project"
version = '0.0.1'
authors = []

[tool.poetry.dependencies]
python = '^3.9'
orjson = "^3.7.11"
pydantic = "^1.9.2"
gevent = "^21.12.0"
flask = "2.2.2"
Flask-SQLAlchemy = "^2.5.1"
Flask-Migrate = "^3.1.0"
uvicorn = "^0.18.3"
psycopg2 = "^2.9.3"
flask-serialize = "^2.1.1"
"backports.pbkdf2" = "^0.1"
PyJWT = "^2.4.0"
redis = "^4.3.4"
pytest-asyncio = "^0.19.0"
flake8 = "^5.0.4"
flasgger = "^0.9.5"
python-dotenv = "^0.21.0"
requests = "^2.28.1"
oauthlib = "^3.2.1"
opentelemetry-api = "^1.13.0"
opentelemetry-sdk = "^1.13.0"
opentelemetry-instrumentation-flask = "^0.34b0"
opentelemetry-exporter-jaeger = "^1.13.0"
Flask-Limiter = "^2.6.3"

[tool.poetry.dev-dependencies]
pytest = '^6.1.2'
aiohttp = "^3.8.1"
redis = '^4.2.2'

[tool.poetry.group.dev.dependencies]
autopep8 = "^1.7.0"
isort = "^5.10.1"

[build-system]
requires = ['poetry-core~=1.0']
build-backend = 'poetry.core.masonry.api'